% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tsne}
\alias{predict.tsne}
\title{predict T-SNE}
\usage{
\method{predict}{tsne}(object, data, k = 5, Y_init = NULL, ...)
}
\arguments{
\item{object}{list; a list from \link{train_stsne}}

\item{data}{matrix; a data want to predict}

\item{k}{numeric; how many neighbor to use (default: 5)}

\item{Y_init}{matrix; If use other method to  set Initial locations of the objects,given the Initial locations in here. If use pca ,keep if NULL (default: NULL).}

\item{...}{Other arguments that can use.}
}
\value{
a dataframe
}
\description{
predict T-SNE
}
