% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sutsne.R
\name{Sutsne}
\alias{Sutsne}
\title{help set weight for the data}
\usage{
Sutsne(Y, X, weight_by_var = TRUE, Y_weight = NULL, X_weight = NULL, ...)
}
\arguments{
\item{Y}{matrix,vector,data.frame; Dependent Variable from other machine learning method}

\item{X}{matrix; Data matrix (each row is an observation, each column is a variable)}

\item{weight_by_var}{logical; use}

\item{Y_weight}{numeric; the weight for Y}

\item{X_weight}{numeric; the weight for X}

\item{...}{Other arguments that can be passed to stsne(except dims,check_duplicates and Y_init )}
}
\value{
List with the following elements:
\item{Y}{Matrix containing the new representations for the objects}
\item{N}{Number of objects}
\item{origD}{Original Dimensionality before TSNE (only when \code{X} is a data matrix)}
\item{perplexity}{See above}
\item{theta}{See above}
\item{costs}{The cost for every object after the final iteration}
\item{itercosts}{The total costs (KL-divergence) for all objects in every 50th + the last iteration}
\item{stop_lying_iter}{Iteration after which the perplexities are no longer exaggerated}
\item{mom_switch_iter}{Iteration after which the final momentum is used}
\item{momentum}{Momentum used in the first part of the optimization}
\item{final_momentum}{Momentum used in the final part of the optimization}
\item{eta}{Learning rate}
\item{exaggeration_factor}{Exaggeration factor used to multiply the P matrix in the first part of the optimization}
\item{weights}{the weight for cost}
\item{r}{a param the can regulation density of point}
}
\description{
help set weight for the data
}
