% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{train_stsne}
\alias{train_stsne}
\title{use pca or other method to set Initial locations ,then the tsne can predict.}
\usage{
train_stsne(data, dims = 2, use_pca = TRUE, Y_init = NULL, ...)
}
\arguments{
\item{data}{matrix; Data matrix (each row is an observation, each column is a variable)}

\item{dims}{integer; Output dimensionality (default: 2)}

\item{use_pca}{logical; If TRUE,will use pca to set Initial locations(default: TRUE)}

\item{Y_init}{matrix; If use other method to  set Initial locations of the objects,given the Initial locations in here. If use pca ,keep if NULL (default: NULL).}

\item{...}{Other arguments that can be passed to stsne(except dims,check_duplicates and Y_init )}
}
\value{
List with the following elements:
\item{Y}{Matrix containing the new representations for the objects}
\item{N}{Number of objects}
\item{origD}{Original Dimensionality before TSNE (only when \code{X} is a data matrix)}
\item{perplexity}{See above}
\item{theta}{See above}
\item{costs}{The cost for every object after the final iteration}
\item{itercosts}{The total costs (KL-divergence) for all objects in every 50th + the last iteration}
\item{stop_lying_iter}{Iteration after which the perplexities are no longer exaggerated}
\item{mom_switch_iter}{Iteration after which the final momentum is used}
\item{momentum}{Momentum used in the first part of the optimization}
\item{final_momentum}{Momentum used in the final part of the optimization}
\item{eta}{Learning rate}
\item{exaggeration_factor}{Exaggeration factor used to multiply the P matrix in the first part of the optimization}
\item{weights}{the weight for cost}
\item{r}{a param the can regulation density of point}
\item{Y_init}{the MDS value to Y_init}
\item{use_pca}{whether use pca to set Initial locations}
\item{pca_coef}{we use pca to measuring MDS, we keep the value for predict}
\item{center}{the data center, predict needs to use it}
}
\description{
use pca or other method to set Initial locations ,then the tsne can predict.
}
